Fixes variable types to properly receive data and drive motor. Fixes scaling equation

+++ L5_Application/ElectricBoardControl.hpp
@@ -12,9 +12,9 @@ public:
     virtual bool init(void);
     virtual bool run(void *param);
 private:
-    const float MAX_DUTY = 8.7;         // % Duty Cycle
-    const float IDLE_DUTY = 7.2;        // % Duty Cycle
-    const float MIN_DUTY = 4.5;
+    const double MAX_DUTY = 8.7;         // % Duty Cycle
+    const double IDLE_DUTY = 7.2;        // % Duty Cycle
+    const double MIN_DUTY = 4.5;
 
     QueueHandle_t commandQueue;
     PWM *motorChannel1;
